Moved the cache to its own file
